feat: update deal ticket button
