+++ libs/deal-ticket/src/components/deal-ticket/deal-ticket-button.tsx
@@ -1,65 +1,27 @@
 import { t } from '@vegaprotocol/react-helpers';
 import type { ButtonVariant } from '@vegaprotocol/ui-toolkit';
-import { ExternalLink } from '@vegaprotocol/ui-toolkit';
-import { Intent } from '@vegaprotocol/ui-toolkit';
-import { Button, Notification } from '@vegaprotocol/ui-toolkit';
-import { useVegaWallet, useVegaWalletDialogStore } from '@vegaprotocol/wallet';
+import { Button } from '@vegaprotocol/ui-toolkit';
+import { useVegaWallet } from '@vegaprotocol/wallet';
 
 interface Props {
   disabled: boolean;
   variant: ButtonVariant;
-  assetSymbol: string;
 }
 
-export const DealTicketButton = ({ disabled, variant, assetSymbol }: Props) => {
+export const DealTicketButton = ({ disabled, variant }: Props) => {
   const { pubKey, isReadOnly } = useVegaWallet();
-  const openVegaWalletDialog = useVegaWalletDialogStore(
-    (store) => store.openVegaWalletDialog
-  );
-  return pubKey && !isReadOnly ? (
+  const isDisabled = Boolean(pubKey) && !isReadOnly ? disabled : true;
+  return (
     <div className="mb-4">
       <Button
         variant={variant}
         fill
         type="submit"
-        disabled={disabled}
+        disabled={isDisabled}
         data-testid="place-order"
       >
         {t('Place order')}
       </Button>
     </div>
-  ) : (
-    <>
-      <Notification
-        intent={Intent.Warning}
-        message={
-          <p className="text-sm pb-2">
-            You need a{' '}
-            <ExternalLink href="https://vega.xyz/wallet">
-              Vega wallet
-            </ExternalLink>{' '}
-            with {assetSymbol} to start trading in this market.
-          </p>
-        }
-        buttonProps={{
-          text: t('Connect wallet'),
-          action: openVegaWalletDialog,
-          dataTestId: 'connect-wallet',
-          size: 'md',
-        }}
-      />
-      <div>
-        <Button
-          variant={variant}
-          fill
-          type="submit"
-          disabled={true}
-          data-testid="place-order"
-          className="my-2"
-        >
-          {t('Place order')}
-        </Button>
-      </div>
-    </>
   );
 };
